rpc/python/pyproject.toml
[tool.poetry]
name = "tiga-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
grpcio-tools = "1.46.0"
grpcio = "^1.59.0"
cachetools = "^5.3.1"
msgpack = "^1.0.7"


[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "default"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
